+++ lib/presentation/home/pages/home_page.dart
@@ -1,8 +1,11 @@
+import 'dart:async';
 import 'package:async_redux/async_redux.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_sficon/flutter_sficon.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'sport_category_page.dart';
+import 'f1_detail_page.dart';
+import 'golf_detail_page.dart';
 import '../../../redux/app_state.dart';
 import '../../../redux/actions/team_actions.dart';
 import '../../../data/models/game.dart';
@@ -18,20 +21,46 @@ import '../widgets/golf_card.dart';
 import '../widgets/tennis_card.dart';
 import '../widgets/rally_card.dart';
 import '../widgets/football_card.dart';
+import '../../widgets/shimmer_loading.dart';
 
-class HomeScreen extends StatelessWidget {
+class HomeScreen extends StatefulWidget {
   const HomeScreen({super.key});
+
+  @override
+  State<HomeScreen> createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends State<HomeScreen> {
+  Timer? _refreshTimer;
+
+  @override
+  void initState() {
+    super.initState();
+    _startRefreshTimer();
+  }
+
+  @override
+  void dispose() {
+    _refreshTimer?.cancel();
+    super.dispose();
+  }
+
+  void _startRefreshTimer() {
+    // Refresh every 3 minutes
+    _refreshTimer = Timer.periodic(const Duration(minutes: 3), (timer) {
+      if (mounted) {
+        StoreProvider.dispatch<AppState>(context, LoadAllGamesAction(showLoading: false));
+      }
+    });
+  }
 
   @override
   Widget build(BuildContext context) {
     return StoreConnector<AppState, _ViewModel>(
       onInit: (store) {
-        if (store.state.games.isEmpty) {
-          store.dispatch(LoadMockGamesAction());
-        }
         store.dispatch(LoadAllGamesAction());
       },
-      vm: () => _Factory(this),
+      vm: () => _Factory(widget),
       builder: (context, vm) {
         return DefaultTabController(
           length: 3,
@@ -81,19 +110,13 @@ class HomeScreen extends StatelessWidget {
                     },
                     color: const Color(0xFFFF6A1A),
                     backgroundColor: const Color(0xFF1A1A1E),
-                    child: vm.isLoading && vm.upcomingGames.isEmpty && vm.liveGames.isEmpty && vm.resultsGames.isEmpty
-                        ? const Center(
-                            child: CircularProgressIndicator(
-                              color: Color(0xFFFF6A1A),
-                            ),
-                          )
-                        : TabBarView(
-                            children: [
-                              _buildGamesList(context, vm.upcomingGames),
-                              _buildGamesList(context, vm.liveGames),
-                              _buildGamesList(context, vm.resultsGames),
-                            ],
-                          ),
+                    child: TabBarView(
+                      children: [
+                        _buildTabContent(context, vm, vm.upcomingGames),
+                        _buildTabContent(context, vm, vm.liveGames),
+                        _buildTabContent(context, vm, vm.resultsGames),
+                      ],
+                    ),
                   ),
                 ),
               ],
@@ -101,6 +124,26 @@ class HomeScreen extends StatelessWidget {
           ),
         );
       },
+    );
+  }
+
+  Widget _buildTabContent(BuildContext context, _ViewModel vm, List<Game> games) {
+    if (vm.isLoading && games.isEmpty) {
+      return _buildShimmerList();
+    }
+    return _buildGamesList(context, games);
+  }
+
+  Widget _buildShimmerList() {
+    return ListView(
+      padding: const EdgeInsets.all(16),
+      children: const [
+        F1CardShimmer(),
+        NBAFootballCardShimmer(),
+        GolfCardShimmer(),
+        TennisCardShimmer(),
+        RallyCardShimmer(),
+      ],
     );
   }
 
@@ -273,62 +316,101 @@ class HomeScreen extends StatelessWidget {
   Widget _buildGameCard(Game game) {
     if (game is F1Game) {
       if (game.status == 'Upcoming') {
-        return F1UpcomingCard(
-          raceName: game.stadium ?? 'TBD Grand Prix',
-          location: game.leagueType ?? 'TBD',
-          raceDate: game.startTime,
-          circuitImage: game.eventImageUrl,
-          broadcastChannel: game.broadcastChannel,
-          raceNumber: game.raceNumber ?? 0,
-          laps: game.laps ?? 0,
-          circuitLayoutUrl: game.circuitLayoutUrl ??
-              'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2b/Circuit_Red_Bull_Ring.svg/1024px-Circuit_Red_Bull_Ring.svg.png',
-          trackLength: game.trackLength ?? '---',
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => F1DetailPage(game: game),
+              ),
+            );
+          },
+          child: Hero(
+            tag: 'f1_${game.id}',
+            child: F1UpcomingCard(
+              raceName: game.stadium ?? 'TBD Grand Prix',
+              location: game.leagueType ?? 'TBD',
+              raceDate: game.startTime,
+              circuitImage: game.eventImageUrl,
+              broadcastChannel: game.broadcastChannel,
+              raceNumber: game.raceNumber ?? 0,
+              laps: game.laps ?? 0,
+              circuitLayoutUrl: game.circuitLayoutUrl ??
+                  'https://upload.wikimedia.org/wikipedia/commons/thumb/2/2b/Circuit_Red_Bull_Ring.svg/1024px-Circuit_Red_Bull_Ring.svg.png',
+              trackLength: game.trackLength ?? '---',
+            ),
+          ),
         );
       } else if (game.status == 'Live') {
-        return F1LiveCard(
-          raceName: game.stadium ?? 'Grand Prix',
-          raceNumber: game.raceNumber ?? 0,
-          circuitImage: game.eventImageUrl,
-          leaderName: game.winnerName ?? 'Unknown',
-          leaderTeam: game.winnerTeam ?? 'TBD',
-          leaderImage: game.winnerImage ??
-              'https://a.espncdn.com/i/teamlogos/f1/500/f1.png',
-          lapInfo: game.winningTime ?? 'LAP --/--',
-          p2Name: game.p2Name,
-          p2Team: game.p2Team,
-          p2Image: game.p2Image,
-          p2Gap: game.p2Gap,
-          p3Name: game.p3Name,
-          p3Team: game.p3Team,
-          p3Image: game.p3Image,
-          p3Gap: game.p3Gap,
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => F1DetailPage(game: game),
+              ),
+            );
+          },
+          child: Hero(
+            tag: 'f1_${game.id}',
+            child: F1LiveCard(
+              raceName: game.stadium ?? 'Grand Prix',
+              raceNumber: game.raceNumber ?? 0,
+              circuitImage: game.eventImageUrl,
+              leaderName: game.winnerName ?? 'Unknown',
+              leaderTeam: game.winnerTeam ?? 'TBD',
+              leaderImage: game.winnerImage ??
+                  'https://a.espncdn.com/i/teamlogos/f1/500/f1.png',
+              lapInfo: game.winningTime ?? 'LAP --/--',
+              p2Name: game.p2Name,
+              p2Team: game.p2Team,
+              p2Image: game.p2Image,
+              p2Gap: game.p2Gap,
+              p3Name: game.p3Name,
+              p3Team: game.p3Team,
+              p3Image: game.p3Image,
+              p3Gap: game.p3Gap,
+            ),
+          ),
         );
       } else {
-        return F1CompletedCard(
-          raceName: game.stadium ?? 'Grand Prix',
-          raceDate: game.startTime,
-          raceNumber: game.raceNumber ?? 0,
-          circuitImage: game.eventImageUrl,
-          winnerName: game.winnerName ?? 'Unknown',
-          winnerTeam: game.winnerTeam ?? 'TBD',
-          winnerLogo: game.winnerImage ??
-              'https://a.espncdn.com/i/teamlogos/f1/500/f1.png',
-          points: game.winnerPoints ?? '0',
-          winnerTotalPoints: game.winnerTotalPoints,
-          time: game.winningTime,
-          p2Name: game.p2Name,
-          p2Team: game.p2Team,
-          p2Image: game.p2Image,
-          p2Points: game.p2Points,
-          p2TotalPoints: game.p2TotalPoints,
-          p2Gap: game.p2Gap,
-          p3Name: game.p3Name,
-          p3Team: game.p3Team,
-          p3Image: game.p3Image,
-          p3Points: game.p3Points,
-          p3TotalPoints: game.p3TotalPoints,
-          p3Gap: game.p3Gap,
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => F1DetailPage(game: game),
+              ),
+            );
+          },
+          child: Hero(
+            tag: 'f1_${game.id}',
+            child: F1CompletedCard(
+              raceName: game.stadium ?? 'Grand Prix',
+              raceDate: game.startTime,
+              raceNumber: game.raceNumber ?? 0,
+              circuitImage: game.eventImageUrl,
+              winnerName: game.winnerName ?? 'Unknown',
+              winnerTeam: game.winnerTeam ?? 'TBD',
+              winnerLogo: game.winnerImage ??
+                  'https://a.espncdn.com/i/teamlogos/f1/500/f1.png',
+              points: game.winnerPoints ?? '0',
+              winnerTotalPoints: game.winnerTotalPoints,
+              time: game.winningTime,
+              p2Name: game.p2Name,
+              p2Team: game.p2Team,
+              p2Image: game.p2Image,
+              p2Points: game.p2Points,
+              p2TotalPoints: game.p2TotalPoints,
+              p2Gap: game.p2Gap,
+              p3Name: game.p3Name,
+              p3Team: game.p3Team,
+              p3Image: game.p3Image,
+              p3Points: game.p3Points,
+              p3TotalPoints: game.p3TotalPoints,
+              p3Gap: game.p3Gap,
+            ),
+          ),
         );
       }
     }
@@ -336,29 +418,68 @@ class HomeScreen extends StatelessWidget {
     if (game is GolfGame) {
       final leaders = game.leaderboard ?? [];
       if (game.status == 'Upcoming') {
-        return GolfUpcomingCard(
-          tournamentName: game.tournamentName ?? 'TBD Tournament',
-          location: game.stadium ?? 'TBD Course',
-          startDate: game.startTime,
-          par: game.par ?? '72',
-          tourType: game.tourType ?? 'PGA Tour',
-          broadcastChannel: game.broadcastChannel,
-          purse: game.purse,
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => GolfDetailPage(game: game),
+              ),
+            );
+          },
+          child: Hero(
+            tag: 'golf_${game.id}',
+            child: GolfUpcomingCard(
+              tournamentName: game.tournamentName ?? 'TBD Tournament',
+              location: game.stadium ?? 'TBD Course',
+              startDate: game.startTime,
+              par: game.par ?? '72',
+              tourType: game.tourType ?? 'PGA Tour',
+              broadcastChannel: game.broadcastChannel,
+              purse: game.purse,
+            ),
+          ),
         );
       } else if (game.status == 'Live') {
-        return GolfLiveCard(game: game);
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => GolfDetailPage(game: game),
+              ),
+            );
+          },
+          child: Hero(
+            tag: 'golf_${game.id}',
+            child: GolfLiveCard(game: game),
+          ),
+        );
       } else {
-        return GolfCompletedCard(
-          tournamentName: game.tournamentName ?? 'TBD Tournament',
-          winnerName: leaders.isNotEmpty ? leaders[0].name : 'Unknown',
-          winnerScore: leaders.isNotEmpty ? leaders[0].score : 'E',
-          winnerImage: leaders.isNotEmpty ? leaders[0].image : null,
-          tourType: game.tourType ?? 'PGA Tour',
-          winnerPurse: game.winnerPurse ?? game.purse,
-          p2Name: leaders.length > 1 ? leaders[1].name : null,
-          p2Score: leaders.length > 1 ? leaders[1].score : null,
-          p3Name: leaders.length > 2 ? leaders[2].name : null,
-          p3Score: leaders.length > 2 ? leaders[2].score : null,
+        return GestureDetector(
+          onTap: () {
+            Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => GolfDetailPage(game: game),
+              ),
+            );
+          },
+          child: Hero(
+            tag: 'golf_${game.id}',
+            child: GolfCompletedCard(
+              tournamentName: game.tournamentName ?? 'TBD Tournament',
+              winnerName: leaders.isNotEmpty ? leaders[0].name : 'Unknown',
+              winnerScore: leaders.isNotEmpty ? leaders[0].score : 'E',
+              winnerImage: leaders.isNotEmpty ? leaders[0].image : null,
+              tourType: game.tourType ?? 'PGA Tour',
+              winnerPurse: game.winnerPurse ?? game.purse,
+              p2Name: leaders.length > 1 ? leaders[1].name : null,
+              p2Score: leaders.length > 1 ? leaders[1].score : null,
+              p3Name: leaders.length > 2 ? leaders[2].name : null,
+              p3Score: leaders.length > 2 ? leaders[2].score : null,
+            ),
+          ),
         );
       }
     }